requirements.txt
annotated-types==0.7.0
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
backoff==2.2.1
boilerpy3==1.0.7
cattrs==23.2.3
certifi==2022.12.7
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
docopt==0.6.2
elastic-transport==8.13.0
elasticsearch==8.13.1
email_validator==2.1.1
Events==0.5
farm-haystack==1.25.5
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
inflect==7.2.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy-imports==0.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
networkx==3.3
num2words==0.5.13
numpy==1.26.4
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
posthog==3.5.0
prompthub-py==4.0.0
pydantic==1.10.15
pydantic_core==2.18.2
Pygments==2.18.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
quantulum3==0.9.1
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-cache==0.9.8
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sseclient-py==1.8.0
starlette==0.37.2
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
tqdm==4.66.4
transformers==4.39.3
typeguard==4.2.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
url-normalize==1.4.3
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
